pyproject.toml
[project]
name = "budget2025-rag"
version = "0.1.0"
description = "Budget 2025 RAG"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "arize-phoenix>=8.24.2",
    "guardrails-ai>=0.5.0",
    "huggingface-hub[cli,inference]>=0.30.2",
    "litellm>=1.66.0",
    "llama-index-callbacks-arize-phoenix>=0.5.0",
    "llama-index-core>=0.12.30",
    "llama-index-embeddings-huggingface>=0.5.3",
    "llama-index-llms-huggingface>=0.5.0",
    "llama-index-llms-huggingface-api>=0.4.1",
    "llama-index-llms-litellm>=0.4.1",
    "llama-index-llms-openai>=0.3.33",
    "llama-index-postprocessor-jinaai-rerank>=0.3.0",
    "llama-index-storage-chat-store-postgres>=0.2.0",
    "llama-index-vector-stores-postgres>=0.4.2",
    "mistralai>=1.6.0",
    "python-dotenv>=1.1.0",
    "streamlit>=1.44.1",
    "streamlit-chat>=0.1.1",
    "tavily-python>=0.5.4",
    "transformers[torch]>=4.51.2",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
    "ipywidgets>=8.1.6",
    "llama-index-utils-workflow>=0.3.1",
]
